+++ tests/test_parser.py
@@ -1,4 +1,5 @@
 from werobot.parser import parse_user_msg
+from werobot.utils import to_text
 
 
 def test_none_message():
@@ -21,7 +22,7 @@ def test_text_message():
     assert message.time == 1348831860
     assert message.type == 'text'
     assert message.content == 'this is a test'
-    assert message.id == 1234567890123456
+    assert message.id == to_text(1234567890123456)
 
 
 def test_image_message():
@@ -40,6 +41,7 @@ def test_image_message():
     assert message.time == 1348831860
     assert message.type == 'image'
     assert message.img == 'this is a url'
+    assert message.id == to_text(1234567890123456)
 
 
 def test_location_message():
@@ -63,6 +65,7 @@ def test_location_message():
     assert message.location == (23.134521, 113.358803)
     assert message.scale == 20
     assert message.label == 'Location'
+    assert message.id == to_text(1234567890123456)
 
 
 def test_link_message():
@@ -85,6 +88,7 @@ def test_link_message():
     assert message.title == 'WeRoBot'
     assert message.description == 'Link to WeRoBot'
     assert message.url == 'https://github.com/whtsky/WeRoBot'
+    assert message.id == to_text(1234567890123456)
 
 
 def test_event_message():